am62px: DSS: Fix display not visible with dss_display_test example.
- dss_display_test examples uses DPI over HDMI to show display output.
The HDMI brigde configuration takes place over I2C and it takes
considerable amount of time for some panels to come out of low power and
put out display. If the frame count that are displayed are very less,
user will not be able to see the display within that time and example
will deinitialise the display also.
- In order to fix this, increase the number of frames to display so that
display is active for a longer time.
- Update the frame count from 300 to 600 (10sec, 60fps)

Fixes: SITSW-6161

Signed-off-by: Soumya Tripathy <s-tripathy@ti.com>