allsocs: dm: Fix Sciclient_pmSetModuleRst_flags doesn't respect flag 0
Sciclient_pmSetModuleRst_flags requires reqFlag as an
argument but when passed 0 as the request flag value
to specify no response is required, it doesn't respect
the flag and returns failure if NACK is received.

Fix this by adding a check for reqFlag not being 0 to
the NACK response check before returning failure.

Fixes: SITSW-6192

Signed-off-by: Sparsh Kumar <sparsh-kumar@ti.com>