am62x/am62ax/am62px: dpl: Update usec range for Timer
Current threshold of Timer period only allows
to setup a maximum timeout of 1 second.

This limit is very less for various usecases,
such as setting up timeout value for LPM
wakeup.

Update the higher threshold for usec for
Timer module to allow setting higher
timeout value for the Timer of 10 seconds.

Fixes: SITSW-6769

Signed-off-by: Sparsh Kumar <sparsh-kumar@ti.com>